src/requirements.txt
aiofiles==23.2.1
flake8==6.1.0
isort==5.12.0
loguru==0.7.2
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
